feature: Display list update progressview only when list is empty refactor: code cleaning
